Make IOTune run throughput tests on parallel shards
We've noticed during experiments that running throughput
workloads with all shards running in parallel, the throughput
tends to drop on raid devices when increasing the number of shards.
The shards sometimes start syncing and due to xfs alignment of files
for the ride array, the shards sometimes end up writing on the same
drive.

Running the throughput workload with shards in parallel brings IOTune
closer to what happens in real workloads, giving it a chance to catch
such cases and not render potentially higher bandwidth numbers than
what we can achieve in practice.

Signed-off-by: Robert Bindar <robert.bindar@scylladb.com>